requirements.txt
telethon
python-dotenv
openai>=1.12.0
python-telegram-bot==13.15
aiogram==3.20.0.post0
aiohttp
beautifulsoup4
requests
Pillow
pytz>=2023.3
